pyproject.toml
[project]
name = "image-sage-mcp"
version = "0.1.0"
description = "MCP server providing Image Sage tool for image analysis via URLs or local files"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Image Sage" }]
dependencies = [
  "mcp>=0.2.0",
  "httpx>=0.27.0",
  "pillow>=10.3.0",
]

[project.optional-dependencies]
vision = [
  "openai>=1.37.0",
  "anthropic>=0.30.0",
  "transformers>=4.42.0",
  "torch>=2.2.0; platform_system != 'Windows'",
]

[project.scripts]
image-sage-mcp = "image_sage_mcp.server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


